Reference

Contents

Reference

Contents

Index

Index

Envelopt.EnveloptLBFGSModel

Envelopt.EnveloptLSR1Model

Envelopt.EnveloptNLPModel

Envelopt.envelopt

Envelopt.EnveloptLBFGSModel — Method

Construct a structured LBFGSModel from an EnveloptNLPModel.

Envelopt.EnveloptLSR1Model — Method

Construct a structured LSR1Model from an EnveloptNLPModel.

Envelopt.EnveloptNLPModel — Type

A structure to represent the problem

minimize f(x) + h_μ (F(x) + μ y)  subject to  c(x) ∈ C,

where h_μ is the Moreau envelope of h with parameter μ > 0.

EnveloptNLPModel(
  model::AbstractNLPModel{T, S},
  F::AbstractNLPModel{T, S},
  h,
  μ::T = 1.0,
) where {T, S}

Arguments

model::AbstractNLPModel{T, S}: a model that represents the problem

minimize_x f(x) subject to c(x) ∈ C.

If model is an NCLModel, model F below must be defined accordingly.

F::AbstractNLPModel{T, S}: a model that represents the mapping F via the feasiblity problem

minimize_x 0 subject to F(x) = 0.

The objective and right-hand side of the constraints are ignored; only the body of the constraints is used. The feasiblity problem above is merely a convenient way to define a mapping F with access to its Jacobian.

If model is an NCLModel (in variables x and r), F must represent a problem of the form

minimize_{x,r} 0 subject to F(x) = 0,

i.e., it has the same number of variables as model, but the constraints are independent of r.

If argument F is omitted, it will be defined implictly as the identity mapping, i.e., F(x) = x, or, in the case of NCLModels, F(x, r) = x.

h: a proper, closed, convex function from ProximalOperators.jl.

Optional arguments

μ::T: the initial penalty parameter (default: 1.0).

Envelopt.envelopt — Method

envelopt(::AbstractNLPModel, args...; kwargs...)
envelopt(::EnveloptNLPModel, args...; kwargs...)

A solver for the nonsmooth regularized problem

minimize f(x) + h(F(x))  subject to  c(x) ∈ C,

where h is a proper, closed, convex function, and f, F and c are smooth. In the first calling form, the AbstractNLPModel represents the problem

minimize f(x) subject to  c(x) ∈ C,

h should be part of args..., and F is implicity set to the identity mapping (F(x) = x).

In the second calling form, the EnveloptNLPModel will have been constructed from a user-defined F.

The Envelopt solver solves a sequence of Moreau envelope approximations of the problem via an augmented Lagrangian method. Each subproblem has the form

minimize f(x) + h_μ (F(x) + μ y)  subject to  c(x) ∈ C,

where h_μ is the Moreau envelope of g with parameter μ > 0, and y is a dual variable estimate. The subproblems are solved with a solver chosen by the user using a quasi-Newton approximation of the Hessian.